# LMS_G03/LMS_G03/bin/Release/netcoreapp3.1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LMS_G03",
    "version": "1.0"
  },
  "paths": {
    "/api/Admin/getuser": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Admin/getuser/{role}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "roleName",
            "type": "string"
          },
          {
            "in": "path",
            "name": "role",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Admin/addnewrole": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "roleName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Admin/viewrole/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Admin/changerole": {
      "post": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ManageUserRolesViewModel"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Admin/removerole": {
      "post": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json",
          "application/xml",
          "text/xml",
          "application/*+xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ManageUserRolesViewModel"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/GoogleDrive/createfile": {
      "post": {
        "tags": [
          "GoogleDrive"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "uploadFile",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "folderName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/GoogleDrive/replacefile": {
      "post": {
        "tags": [
          "GoogleDrive"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "oldFileId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "parentsId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "FileName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ManageUserRolesViewModel": {
      "type": "object",
      "properties": {
        "roleId": {
          "type": "string"
        },
        "roleName": {
          "type": "string"
        },
        "selected": {
          "type": "boolean"
        }
      }
    },
...
  }
}